+++ client/src/components/kanban/board.js
@@ -1,0 +1,8 @@
+export default function Board(props) {
+    const { className: progressStyle } = props;
+    return (
+        <div className={`kanban-board ${progressStyle}`}>
+            This is a kanban board
+        </div>
+    );
+}
