+++ client/src/components/task/taskViewSmall.js
@@ -1,0 +1,1 @@
+export default function TaskViewSmall() {}
