--- client/src/overview-components/overview.js
@@ -1,10 +1,0 @@
-import TasksList from "../task-components/tasksList";
-
-export default function Overview() {
-    return (
-        <div>
-            <h1>Hello from the Overview Component</h1>
-            <TasksList />
-        </div>
-    );
-}
